Update PHS_NASSS_Cardiac_WebPlotDigitizerSaveData.R
There were problems with the cumsum() values in the Pubs graphs. This code snippet tests the code and checks to see why the cumulative results graph differ from the cumulative sums in the .summary data frames here.

Conclusion: the graphs don't use the fully modelled and interpolated data. They use the data straight from the reports that has gaps - hence, the sums are lower.